restruction.add forcecontrol.cpp, forcecontrol.h
