+++ CMakeLists.txt
@@ -17,8 +17,7 @@ include_directories(${aris_INCLUDE_DIRS})
 
 message(${aris_DIR})
 
-
 # user program started...
-set(SOURCE_FILES src/main.cpp src/rokae.cpp src/iir.cpp src/rokae.h src/iir.h)
+set(SOURCE_FILES src/main.cpp src/kaanh.cpp src/iir.cpp src/forcecontrol.cpp src/kaanh.h src/iir.h src/forcecontrol.h)
 add_executable(kaanh ${SOURCE_FILES})
 target_link_libraries(kaanh ${aris_LIBRARIES})
